requirements.txt
fastapi
uvicorn
pyhumps
pandas
